# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "WhatsApp/NFS-e API",
        "description": "API de integração entre agendamentos e providers WhatsApp/NFS-e.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Servidor atual"
        }
    ],
    "paths": {
...
        "/api/v1/messages": {
            "post": {
                "tags": [
                    "Messages"
                ],
                "summary": "Enfileira uma notificação",
                "description": "Autenticação HMAC entre sistemas. `establishment_external_id` é opcional hoje (número central) e passa a identificar o estabelecimento na fase de número próprio por tenant.",
                "operationId": "a82ec5d449b34296833ea2348fd5805d",
                "parameters": [
                    {
                        "name": "X-Client-Id",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "agendamentos"
                        }
                    },
                    {
                        "name": "X-Timestamp",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1788000000
                        }
                    },
                    {
                        "name": "X-Nonce",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "a1b2c3d4e5f6"
                        }
                    },
                    {
                        "name": "X-Signature",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "sha256=..."
                        }
                    },
                    {
                        "name": "Idempotency-Key",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "event",
                                    "recipient"
                                ],
                                "properties": {
                                    "event": {
                                        "type": "string",
                                        "example": "appointment.created"
                                    },
                                    "establishment_external_id": {
                                        "description": "Opcional; identifica o estabelecimento de origem",
                                        "type": "string",
                                        "example": "est-001",
                                        "nullable": true
                                    },
                                    "recipient": {
                                        "properties": {
                                            "role": {
                                                "type": "string",
                                                "example": "customer"
                                            },
                                            "phone": {
                                                "type": "string",
                                                "example": "+5511999999999"
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "template": {
                                        "properties": {
                                            "name": {
                                                "type": "string",
                                                "example": "lembrete_agendamento_v1"
                                            },
                                            "language": {
                                                "type": "string",
                                                "example": "pt_BR"
                                            },
                                            "parameters": {
                                                "type": "array",
                                                "items": {
                                                    "properties": {
                                                        "name": {
                                                            "type": "string"
                                                        },
                                                        "value": {
                                                            "type": "string"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            }
                                        },
                                        "type": "object",
                                        "nullable": true
                                    },
                                    "text": {
                                        "properties": {
                                            "body": {
                                                "type": "string",
                                                "example": "Bom dia! Confirmamos seu agendamento."
                                            }
                                        },
                                        "type": "object",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Requisicao duplicada idempotente"
                    },
                    "202": {
                        "description": "Aceito para processamento"
                    },
                    "401": {
                        "description": "Assinatura inválida"
                    },
                    "422": {
                        "description": "Dados inválidos"
                    }
                },
                "security": [
                    {
                        "hmac": []
                    }
                ]
            }
        },
        "/api/v1/messages/{message}": {
            "get": {
                "tags": [
                    "Messages"
                ],
                "summary": "Consulta o status de uma notificação",
                "operationId": "6aba0ca4020fa07211c59c97f567e35b",
                "responses": {
                    "200": {
                        "description": "Status da mensagem"
                    },
                    "401": {
                        "description": "Assinatura inválida"
                    },
                    "404": {
                        "description": "Não encontrada"
                    }
                },
                "security": [
                    {
                        "hmac": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "hmac": {
                "type": "apiKey",
                "name": "X-Signature",
                "in": "header"
            }
        }
    },
...
}